+++ spring-core/03-bean-scopes/src/main/resources/beans.xml
@@ -4,7 +4,7 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
            https://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="emailService" class="dangminhphuc.dev.demo.beanscopes.EmailNotificationService" scope="singleton"/>
-    <bean id="telegramService" class="dangminhphuc.dev.demo.beanscopes.TelegramNotificationService" scope="prototype"/>
+    <bean id="emailService" class="dangminhphuc.dev.demo.beanscopes.singleton.EmailNotificationService" scope="singleton"/>
+    <bean id="telegramService" class="dangminhphuc.dev.demo.beanscopes.prototype.TelegramNotificationService" scope="prototype"/>
 
 </beans>
